Rename components within apps
