+++ src/applications/appeals/995/config/form.js
@@ -10,7 +10,7 @@ import IntroductionPage from '../containers/IntroductionPage';
 import ConfirmationPage from '../containers/ConfirmationPage';
 import GetFormHelp from '../content/GetFormHelp';
 
-import AddIssue from '../components/AddIssue';
+import AddContestableIssue from '../components/AddContestableIssue';
 import PrimaryPhone from '../components/PrimaryPhone';
 import PrimaryPhoneReview from '../components/PrimaryPhoneReview';
 import EvidenceVaRecords from '../components/EvidenceVaRecords';
@@ -151,7 +151,7 @@ const formConfig = {
           title: 'Add issues for review',
           path: ADD_ISSUE_PATH,
           depends: () => false, // accessed from contestable issues
-          CustomPage: AddIssue,
+          CustomPage: AddContestableIssue,
           CustomPageReview: null,
           uiSchema: {},
           schema: blankSchema,
